+++ src/resources/views/pages/home.blade.php
@@ -71,6 +71,7 @@
                         <li>Added reusable Revision and Revisionable trait for tracking of changes in any model. (issue #138)</li>
                         <li>Added UI for viewing Revision History for Orders.</li>
                         <li>Allow User to change their password and names (issue #159)</li>
+                        <li>Create new Order with product quantities (issue #165)</li>
                     </ol>
                     <h4>Enhancements:</h4>
                     <ol>